changing the data type of one field

+++ DDL_DML/compras_inventario.sql
@@ -12,7 +12,7 @@ CREATE TABLE proveedores(
 CREATE TABLE Facturas(
     numFact SMALLINT,
     fecha DATE,
-    IVA int,
+    IVA DECIMAL(8,2),
     provID VARCHAR(100)
 );
